Use a fork of log-update for displaying output (#282)
`log-update` module wraps output on every render, which slows down performance
of Ink apps. Ink doesn't need that and some other functionality, because it
handles wrapping itself, so `log-update` is essentially performing the same
work twice.

This change adds a forked and stripped down version of `log-update` with the
plan to merge the changes upstream in the future.