Update only changed props and styles of the node
Before this change, Ink updated all props (and styles) even if only
one was changed. This was especially painful for Yoga styles,
because applyStyles() function re-applied all styles for each node.

This change introduces prop diffing to Ink's reconciler config,
which means that only changed props will be updated in the nodes.
In addition, there's a special handling for styling props to
ensure only new, changed and deleted styles are going to be handled.
Styles which remained unchanged aren't even going to make it to
applyStyles() function.