Ignore Ctrl+C in useInput hook
Ctrl+C exits Ink app by default, but if it's not ignored in useInput,
chances are it's going to trigger a state update after unmount.
This results in a warning from React:

Warning: Can't perform a React state update on an unmounted component.
This is a no-op, but it indicates a memory leak in your application.
To fix, cancel all subscriptions and asynchronous tasks in a useEffect cleanup function.