Remeasure text dimensions when text node is changed
Previously remeasuring was triggered when content of the text node
was changed. For example:

<Text>abc</Text> => <Text>abcx</Text>

But when there was a new nested <Text> node added or removed,
remeasuring wasn't triggered, because "abc" text didn't change.

<Text>abc</Text> => <Text>abc<Text>x</Text></Text>

This change ensures that whenever <Text> node or its children
are mutated, Ink finds the closest Yoga node and marks it as
dirty to trigger remeasuring on next render.